Use UTC times for failure times etc.
This fixes problems with negative downtime (e.g. when sending monitors
to a remote logger which is in a TZ behind us (GMT -> EST)), but means
all times reported in webpages etc will be UTC not local time for
monitor. That should be fixed to either use local TZ or include TZ in
times so we know when it means.