JSON serialise compiled re objects safely
This commit fixes an issue reported by a user where
JSONEncoder.default() was called with too many params, and in the
process of fixing that revealed an issue where compiled re objects (as
generated by MonitorHost) can't be serialised.

Address it by serialising it to None, and having MonitorHost
re-initialise it if needed (although generally run_test() isn't called
on objects received over the network).

Fixes #187.