Resort joblist before execution
Make sure compound monitors are pushed to the end of the list.

Not 100% sure this is the right answer still, might need to split the
joblist into two and run one then the other each loop, otherwise a
compound monitor's dependency could be executing at the same time as the
compound monitor checks it.